+++ src/main/java/com/jof/springmvc/service/FriendService.java
@@ -1,6 +1,5 @@
 package com.jof.springmvc.service;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import com.jof.springmvc.model.Friend;
@@ -8,13 +7,13 @@ import com.jof.springmvc.model.User;
 
 public interface FriendService {
 	
-	Friend findFriendshipByIds(BigInteger id_alpha, BigInteger id_omega);
+	Friend findFriendshipByIds(Long id_alpha, Long id_omega);
 
 	void saveFriend(Friend friend);
 	
 	void updateFriend(Friend friend);
 
-	void deleteByIds(BigInteger alpha_id, BigInteger omega_id);
+	void deleteByIds(Long alpha_id, Long omega_id);
 
 	// returns omega ids
 	List<Friend> findAllFriends(User user);
